requirement.txt
cmake
dlib
face-recognition
cvzone
opencv-python



import pickle
import cv2
import os
import face_recognition  
import numpy as np
import cvzone
from pymongo import MongoClient

# MongoDB connection
client = MongoClient("mongodb://localhost:27017/")
db = client["project-FR"]
students_col = db["students"]
profile_col = db["Studentimg"]
print("MongoDB Connected Successfully ✅")

cap = cv2.VideoCapture(0)

folderModePath = 'resources/modes'
modePathList = os.listdir(folderModePath)
imgModeList = [cv2.imread(os.path.join(folderModePath, p)) for p in modePathList]

# Camera settings
CAM_WIDTH, CAM_HEIGHT = 605, 430
X_OFFSET, Y_OFFSET = 53, 167

# Mode settings
MODE_WIDTH, MODE_HEIGHT = 464, 590
MODE_X, MODE_Y = 748, 60

# Load encodings
print("Loading Encoded File....")
with open("EncodeFile.p", "rb") as file:
    encodeListKnown, studentIds = pickle.load(file)

print(studentIds)
print("Encode File Loaded")

modeType = 0
counter = 0
id = None
studentInfo = None
currentStudentImg = None
last_id = None

while True:
    success, img = cap.read()
    if not success:
        break

    # 🔹 Fresh background every frame
    imgBackground = cv2.imread("resources/background.jpeg").copy()

    imgS = cv2.resize(img, (0, 0), None, 0.25, 0.25)
    imgS = cv2.cvtColor(imgS, cv2.COLOR_BGR2RGB)

    faceCurFrame = face_recognition.face_locations(imgS)
    encodeCurFrame = face_recognition.face_encodings(imgS, faceCurFrame)

    img = cv2.resize(img, (CAM_WIDTH, CAM_HEIGHT))
    imgBackground[Y_OFFSET:Y_OFFSET + CAM_HEIGHT,
                  X_OFFSET:X_OFFSET + CAM_WIDTH] = img

    imgMode = cv2.resize(imgModeList[modeType], (MODE_WIDTH, MODE_HEIGHT))
    imgBackground[MODE_Y:MODE_Y + MODE_HEIGHT,
                  MODE_X:MODE_X + MODE_WIDTH] = imgMode

    if faceCurFrame:
        for encodeFace, faceLoc in zip(encodeCurFrame, faceCurFrame):
            matches = face_recognition.compare_faces(encodeListKnown, encodeFace)
            faceDis = face_recognition.face_distance(encodeListKnown, encodeFace)

            matchIndex = np.argmin(faceDis)

            if matches[matchIndex]:
                id = studentIds[matchIndex]

                y1, x2, y2, x1 = faceLoc
                y1, x2, y2, x1 = y1*4, x2*4, y2*4, x1*4
                bbox = 55+x1, 162+y1, x2-x1, y2-y1
                imgBackground = cvzone.cornerRect(imgBackground, bbox, rt=0)

                if id != last_id:
                    studentInfo = students_col.find_one({"_id": id})
                    student_profile = profile_col.find_one({"student_id": str(id)})

                    if student_profile:
                        imgpath = student_profile.get("image_path")
                        if imgpath and os.path.exists(imgpath):
                            currentStudentImg = cv2.imread(imgpath)
                            currentStudentImg = cv2.resize(currentStudentImg, (240, 192))
                        else:
                            currentStudentImg = np.zeros((240,192,3), dtype=np.uint8)

                    last_id = id
                    counter = 1
                    modeType = 1

    # 🔹 Always draw student image if available
    if currentStudentImg is not None:
        imgBackground[175:175+192, 859:859+240] = currentStudentImg

    # 🔹 Text rendering
    if studentInfo:
        cv2.putText(imgBackground, str(studentInfo.get('total_attandance', 0)),
                    (810,135), cv2.FONT_HERSHEY_COMPLEX, 1, (255,255,255), 1)

        cv2.putText(imgBackground, studentInfo.get('name', ''),
                    (1006,540), cv2.FONT_HERSHEY_COMPLEX, 1, (255,255,255), 1)

        cv2.putText(imgBackground, str(id),
                    (1000,448), cv2.FONT_HERSHEY_COMPLEX, 0.5, (255,255,255), 1)

        cv2.putText(imgBackground, studentInfo.get('major', ''),
                    (1000,498), cv2.FONT_HERSHEY_COMPLEX, 0.5, (255,255,255), 1)

        cv2.putText(imgBackground, studentInfo.get('standing', ''),
                    (880,578), cv2.FONT_HERSHEY_COMPLEX, 0.6, (100,100,100), 1)

        cv2.putText(imgBackground, str(studentInfo.get('year', '')),
                    (990,578), cv2.FONT_HERSHEY_COMPLEX, 0.6, (100,100,100), 1)

        cv2.putText(imgBackground, str(studentInfo.get('starting_year', '')),
                    (1105,578), cv2.FONT_HERSHEY_COMPLEX, 0.6, (100,100,100), 1)
        
        (w,h), _ =cv2.getTextSize(studentInfo['name'],cv2.FONT_HERSHEY_COMPLEX, 1, 1)
        offset = (414-w)//2
        cv2.putText(imgBackground ,str(studentInfo['name']),(775 + offset ,410),
                    cv2.FONT_HERSHEY_COMPLEX,1,(50,50,50),1)

    cv2.imshow("Face Attendance", imgBackground)

    if cv2.waitKey(10) & 0xFF == 27:
        break

cap.release()
cv2.destroyAllWindows()


